requirements/base.txt
django>=4.1.7
django-environ>=0.10.0
setuptools>=69.1.1
djangorestframework>=3.14.0
django-cors-headers>=3.14.0
django-filter>=22.1
django-autoslug>=1.9.8
drf-yasg>=1.21.5
Pillow>=9.4.0
argon2-cffi>=21.3.0
pytz>=2022.7.1
django-elasticsearch-dsl>=7.3.0
django-elasticsearch-dsl-drf>=0.22.5